+++ packages/expo/src/winter/runtime.native.ts
@@ -28,7 +28,5 @@ install('URL', () => require('./url').URL);
 install('URLSearchParams', () => require('./url').URLSearchParams);
 
 // Polyfill async iterator symbol for Hermes.
-if (!Symbol.asyncIterator) {
-  // @ts-expect-error: readonly property only applies when the engine supports it
-  Symbol.asyncIterator = Symbol.for('Symbol.asyncIterator');
-}
+// @ts-expect-error: readonly property only applies when the engine supports it
+Symbol.asyncIterator ??= Symbol.for('Symbol.asyncIterator');